Add files via upload
Our project is based on a problem statement that how can we solve the problems faced by the farmers related to soil testing.

Agrisaral is a farmer-friendly platform that provides personalized soil testing, tailored crop and fertilizer recommendations, and valuable agricultural knowledge, all in a language farmers understand. Our goal is to make farming easier, more productive, and more profitable.